+++ src/components/detail-cast-list/CastList.jsx
@@ -16,7 +16,7 @@ const CastList = (props) => {
     const getCredits = async () => {
       const response = await tmdbApi.credits(category, props.id);
       console.log(response);
-      setCasts(response.cast.slice(0, 5));
+      setCasts(response.cast.slice(0, 6));
     };
     getCredits();
   }, [category, props.id]);
