+++ src/components/movie-card/MovieCard.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./movieCard.scss";
 
 import { Link } from "react-router-dom";
-import Button from "../button/Button";
+import Button from "../../common/button/Button";
 
 import { category } from "../../api/tmdbApi";
 import apiConfig from "../../api/apiConfig";
